crawl_dantri/requirements.txt
requests
tqdm
matplotlib
python-slugify
beautifulsoup4
lxml
streamlit
fastapi
uvicorn[standard]
pandas
altair
aiofiles
aiohttp
watchdog
opencv-python
playwright
